refactor: rewrite and organize utils to decrypt, resolve references, sync etc
