+++ backend/api/utils/secrets.py
@@ -1,7 +1,50 @@
+import re
 from django.db import transaction
+from django.apps import apps
 
-from api.models import Environment, SecretFolder, Secret
-from api.utils.syncing.secrets import compute_key_digest
+# from api.models import SecretFolder, Secret, ServerEnvironmentKey
+
+from api.utils.crypto import (
+    blake2b_digest,
+    decrypt_asymmetric,
+    env_keypair,
+    get_server_keypair,
+)
+
+
+def get_environment_keys(environment_id):
+    """
+    Returns a tuple of environment public and private keys.
+
+    Args:
+        environment_id (str): The ID of the environment to get the keys for.
+
+    Returns:
+        env_pubkey, env_privkey (tuple): A tuple containing the environment's public key and private key.
+    """
+    ServerEnvironmentKey = apps.get_model("api", "ServerEnvironmentKey")
+
+    server_env_key = ServerEnvironmentKey.objects.get(environment_id=environment_id)
+
+    pk, sk = get_server_keypair()
+
+    env_seed = decrypt_asymmetric(server_env_key.wrapped_seed, sk.hex(), pk.hex())
+
+    return env_keypair(env_seed)
+
+
+def compute_key_digest(key, environment_id):
+    ServerEnvironmentKey = apps.get_model("api", "ServerEnvironmentKey")
+
+    server_env_key = ServerEnvironmentKey.objects.get(environment_id=environment_id)
+
+    pk, sk = get_server_keypair()
+
+    salt = decrypt_asymmetric(server_env_key.wrapped_salt, sk.hex(), pk.hex())
+
+    key_digest = blake2b_digest(key.upper(), salt)
+
+    return key_digest
 
 
 def create_environment_folder_structure(complete_path, environment_id):
@@ -16,6 +59,9 @@ def create_environment_folder_structure(complete_path, environment_id):
     - SecretFolder: The last `SecretFolder` instance created or retrieved, representing the deepest
                     level in the provided path structure.
     """
+    Environment = apps.get_model("api", "Environment")
+    SecretFolder = apps.get_model("api", "SecretFolder")
+
     # Ensure the path_segments list does not include empty segments caused by leading or trailing slashes
     path_segments = [segment for segment in complete_path.split("/") if segment]
 
@@ -90,6 +136,8 @@ def check_for_duplicates_blind(secrets, environment):
     Returns:
         bool: True if a duplicate is found, False otherwise.
     """
+    Secret = apps.get_model("api", "Secret")
+
     processed_secrets = set()  # Set to store processed secrets
 
     for secret in secrets:
@@ -120,3 +168,133 @@ def check_for_duplicates_blind(secrets, environment):
         processed_secrets.add((path, secret["keyDigest"]))
 
     return False  # No duplicates found
+
+
+def decompose_path_and_key(composed_key):
+    # Determine the path and key name from the rest of the reference.
+    last_slash_index = composed_key.rfind("/")
+    if last_slash_index != -1:
+        path = composed_key[:last_slash_index]
+        key_name = composed_key[last_slash_index + 1 :]
+    else:
+        path = "/"
+        key_name = composed_key
+
+    return normalize_path_string(path), key_name
+
+
+def decrypt_secret_value(secret):
+    """
+    Decrypts the given secret's value and resolves all references.
+
+
+    Args:
+        secret (Secret): The secret instance to decrypt.
+
+
+    Returns:
+        key, value (tuple): A tuple containing the target secret key and value.
+        OR
+        [(key, value)]: A list of tuples containing all secrets' keys and values
+    """
+    Secret = apps.get_model("api", "Secret")
+    Environment = apps.get_model("api", "Environment")
+    ServerEnvironmentKey = apps.get_model("api", "ServerEnvironmentKey")
+
+    cross_env_pattern = re.compile(r"\$\{(.+?)\.(.+?)\}")
+    local_ref_pattern = re.compile(r"\$\{([^.]+?)\}")
+
+    pk, sk = get_server_keypair()
+
+    server_env_key = ServerEnvironmentKey.objects.get(
+        environment_id=secret.environment.id
+    )
+
+    app = server_env_key.environment.app
+
+    env_seed = decrypt_asymmetric(server_env_key.wrapped_seed, sk.hex(), pk.hex())
+
+    env_salt = decrypt_asymmetric(server_env_key.wrapped_salt, sk.hex(), pk.hex())
+
+    env_pubkey, env_privkey = env_keypair(env_seed)
+
+    value = decrypt_asymmetric(secret.value, env_privkey, env_pubkey)
+
+    cross_env_matches = re.findall(cross_env_pattern, value)
+
+    for ref_env, ref_key in cross_env_matches:
+
+        try:
+            path, key_name = decompose_path_and_key(ref_key)
+
+            referenced_environment = Environment.objects.get(
+                name__iexact=ref_env, app=app
+            )
+            referenced_environment_key = ServerEnvironmentKey.objects.get(
+                environment_id=referenced_environment.id
+            )
+            seed = decrypt_asymmetric(
+                referenced_environment_key.wrapped_seed, sk.hex(), pk.hex()
+            )
+            salt = decrypt_asymmetric(
+                referenced_environment_key.wrapped_salt, sk.hex(), pk.hex()
+            )
+
+            key_digest = blake2b_digest(key_name, salt)
+
+            referenced_env_pubkey, referenced_env_privkey = env_keypair(seed)
+
+            referenced_secret = Secret.objects.get(
+                environment=referenced_environment,
+                path=path,
+                key_digest=key_digest,
+                deleted_at=None,
+            )
+
+            referenced_secret_value = decrypt_asymmetric(
+                referenced_secret.value,
+                referenced_env_privkey,
+                referenced_env_pubkey,
+            )
+
+            value = value.replace(f"${{{ref_env}.{ref_key}}}", referenced_secret_value)
+        except:
+            print(
+                f"Warning: The referenced environment or key either does not exist or the server does not have access to it."
+            )
+            pass
+
+    # Handle local references
+    local_ref_matches = re.findall(local_ref_pattern, value)
+
+    for ref_key in local_ref_matches:
+
+        try:
+            path, key_name = decompose_path_and_key(ref_key)
+
+            key_digest = blake2b_digest(key_name, env_salt)
+
+            referenced_secret = Secret.objects.get(
+                environment=secret.environment,
+                path=path,
+                key_digest=key_digest,
+                deleted_at=None,
+            )
+
+            referenced_secret_value = decrypt_asymmetric(
+                referenced_secret.value,
+                env_privkey,
+                env_pubkey,
+            )
+
+            value = value.replace(
+                f"${{{ref_key}}}",
+                referenced_secret_value,
+            )
+        except:
+            print(
+                f"Warning: The referenced environment or key either does not exist or the server does not have access to it."
+            )
+            pass
+
+    return value
